requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.24.0

# Metrics and Evaluation
scikit-image>=0.20.0  # For SSIM calculation

# Training Utilities
tensorboard>=2.12.0
tqdm>=4.65.0

# Visualization (optional)
matplotlib>=3.7.0
